requirements.txt
openai>=1.12.0
ollama
google-genai
